+++ app/src/main/java/com/example/passwordreminder/MainActivity.java
@@ -1,22 +1,26 @@
 package com.example.passwordreminder;
-
+import androidx.appcompat.app.AlertDialog;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.WindowManager;
 import android.widget.Button;
 import android.widget.Toast;
-
-import androidx.appcompat.app.AppCompatActivity;
+import android.text.Editable;
+import android.text.TextWatcher;
+import android.widget.EditText;
 import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-public class MainActivity extends AppCompatActivity {
+public class MainActivity extends BaseSecureActivity {
 
     private VaultViewModel vm;
+    private androidx.lifecycle.LiveData<java.util.List<Credential>> currentSource;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+
+
 
 // ✅ Screenshot + screen record engelle
         getWindow().setFlags(
@@ -56,9 +60,18 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onDelete(Credential c) {
                 if (!isAdmin) return;
-                vm.getRepo().deleteCredential(c);
-                Toast.makeText(MainActivity.this, "Silindi", Toast.LENGTH_SHORT).show();
+
+                new AlertDialog.Builder(MainActivity.this)
+                        .setTitle("Silme Onayı")
+                        .setMessage("Bu kaydı silmek istediğinize emin misiniz?\n\n" + c.title+"\n"+c.username)
+                        .setPositiveButton("Evet, Sil", (dialog, which) -> {
+                            vm.getRepo().deleteCredential(c);
+                            Toast.makeText(MainActivity.this, "Silindi", Toast.LENGTH_SHORT).show();
+                        })
+                        .setNegativeButton("İptal", (dialog, which) -> dialog.dismiss())
+                        .show();
             }
+
         }, isAdmin);
 
         rv.setLayoutManager(new LinearLayoutManager(this));
@@ -70,5 +83,39 @@ public class MainActivity extends AppCompatActivity {
             if (!isAdmin) return;
             startActivity(new Intent(MainActivity.this, EditCredentialActivity.class));
         });
+
+
+        EditText etSearch = findViewById(R.id.etSearch);
+
+// İlk kaynak: tüm liste
+        currentSource = vm.getCredentials();
+        currentSource.observe(this, adapter::submit);
+
+        etSearch.addTextChangedListener(new TextWatcher() {
+            @Override public void beforeTextChanged(CharSequence s, int start, int count, int after) {}
+            @Override public void afterTextChanged(Editable s) {}
+
+            @Override
+            public void onTextChanged(CharSequence s, int start, int before, int count) {
+                String q = s.toString().trim();
+
+                // Eski observer'ı kaldır
+                if (currentSource != null) {
+                    currentSource.removeObservers(MainActivity.this);
+                }
+
+                // Yeni kaynağı seç
+                if (q.isEmpty()) {
+                    currentSource = vm.getCredentials();
+                } else {
+                    currentSource = vm.search(q);
+                }
+
+                // Yeni kaynağı bağla
+                currentSource.observe(MainActivity.this, adapter::submit);
+            }
+        });
+
+
     }
 }
